Add backend functionality and refactor
The function for determining if a click lies in the selected
character's box now queries a database in the backend. The time taken
to find all the characters is now calculated and can be stored in a
database in the backend. The app has been split into separate modules
for each top level component of the app.